requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
frozenlist==1.6.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
loader==2017.9.11
Logbook==1.8.1
multidict==6.4.3
pendulum==3.1.0
propcache==0.3.1
pyaes==1.6.1
pyasn1==0.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==22.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
Telethon==1.39.0
termcc==2019.8.2
terminaltables==3.1.10
TgCrypto==1.2.5
typing_extensions==4.13.2
tzdata==2025.2
worker==2018.4.1
yarl==1.20.0
